backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Keiran Jensen <keiranjensen@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.14.0"
django = "^5.2.7"
psycopg2-binary = "^2.9.11"
djangorestframework = "^3.16.1"
gunicorn = "^23.0.0"
django-environ = "^0.12.0"
djangorestframework-camel-case = "^1.4.2"
django-redis = "^6.0.0"
django-allauth = {extras = ["headless-spec"], version = "^65.14.3"}
pyyaml = "^6.0.3"

[tool.poetry.group.types.dependencies]
types-gunicorn = "^23.0.0.20251031"
types-pyyaml = "^6.0.12.20250915"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytz (>=2025.2,<2026.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "ruff (>=0.14.3,<0.15.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "django-stubs[compatible-mypy] (>=5.2.7,<6.0.0)",
    "djangorestframework-stubs[compatible-mypy] (>=3.16.5,<4.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-xdist (>=3.8.0,<4.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)"
]
